+++ prettyprinter/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Current: 0.2
 
-- Add Functor instances for the Doc types (they reannotate a document)
+- Add Foldable/Traversable instances for `SimpleDocTree`, `SimpleDocStream`
+- Add Functor instances for `Doc`, `SimpleDocTree`, `SimpleDocStream`
 - Add the simplified renderers `renderSimplyDecorated` and
   `renderSimplyDecoratedA` to the tree and stack renderer modules
 - Add a changelog :-)